+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * aes/aes-gcm: add x86/x86-64 optimise backend: AES-NI for the `aes` module and PCLMULQDQ for the GHASH authentication
+* sha2: improve x86 backends (better scheduling, shani, misc optimisations, code cleanup)
 
 # 0.6.5